switch DataModel serialization to JSON
Don't just write out Java objects, as this is not very friendly to
edit and it's prone to break on incompatible updates.  "Plain" JSON
text files should be easier to maintain.

THIS IS AN INCOMPATIBLE CHANGE AND IRREVOCABLY PREVENTS ALL PREVIOUSLY
SAVED DATA MODELS TO BE READ IN AGAIN.  As there currently is only one
application using this library that I know of and this application has
only one user that I know of and that user does not care for old saved
data, I don't care about being incompatible.  Saves me the hassle of
writing a converter. (This should be rather easy, though, if anybody
should ever need this.  Just take care to implement the
reading-from-a-file-part on cgarbs-javalib v0.2.0 to be compatible.)